Major Command Update
Changelog :
- Repaired old commands
- Hide a little bit DevTools Folder
- Commented 99% of commands
- Cleaned database folder

Upcoming :
- Finishing comming command Info + other files
- Add Alias search on Help command
- Clean command Info

+++ commands/DevTools/mp.js
@@ -1,6 +1,6 @@
-const { SlashCommandBuilder } = require('discord.js');
 const { admin } = require('../../settings.json')
 
+// Create the json script for the help command
 module.exports.help = {
     name : "mp",
     description: 'Private Message anyone!',
@@ -9,48 +9,37 @@ module.exports.help = {
     parameters: 'none'
 };
 
+// Create a the run script for the command
 module.exports.execute = async (client, message, args) => {
+    // Check if the author is the admin
     if (!message.author.id === admin) return;
 
+    // Check if the author has tagged someone
     switch (args[0]) {
         case undefined:
         case null:
+            // If not, send a message
             message.reply({content: 'You have to tag or Id someone !', allowedMentions: { repliedUser: false }});
             break;
         default:
-            if (message.mentions.users.first()) {
+            // If yes, send a message to the tagged user and verify if there is some text
+            if (message.mentions.users.first() && args.length > 1) {
+                // Get the user tag object
                 const getMentionTag = message.mentions.users.first();
 
+                // Delete the original message
                 setTimeout(() => {message.delete()}, 500)
 
+                // Send the message to the tagged user
                 getMentionTag.send(args.slice(1).join(' '));
+
             } else {
-                message.reply({content: 'You have to tag or Id someone !', allowedMentions: { repliedUser: false }});
+                // If not, send a message
+                message.reply({content:
+                    'You have to tag or Id someone with some text !',
+                    allowedMentions: { repliedUser: false }
+                });
             }
             break;
     }
-};
-
-module.exports.data = new SlashCommandBuilder()
-    .setName(module.exports.help.name)
-    .setDescription(module.exports.help.description)
-    .addUserOption(option =>
-        option
-            .setName('user')
-            .setDescription('The user to send the message to')
-            .setRequired(true))
-    .addStringOption(option =>
-        option
-            .setName('message')
-            .setDescription('The message to send')
-            .setRequired(true))
-    .setDMPermission(false)
-
-module.exports.run = async (client, interaction) => {
-    if (!(interaction.member?.user.id ?? interaction.user.id) === admin) return;
-
-    const userMention = interaction.options.getUser('user');
-    const message = interaction.options.getString('message');
-
-    userMention.send(message);
 };
